+++ src/controllers/SubscriptionsController.php
@@ -43,6 +43,19 @@ class SubscriptionsController extends BaseController
         return $this->renderTemplate('commerce/subscriptions/_index');
     }
 
+
+    /**
+     * @param Subscription $subscription
+     * @throws ForbiddenHttpException
+     * @since 3.4.8
+     */
+    protected function enforceEditSubscriptionPermissions(Subscription $subscription): void
+    {
+        if (!$subscription->getIsEditable()){
+            throw new ForbiddenHttpException('User is not permitted to edit this subscription');
+        }
+    }
+
     /**
      * @param int|null $subscriptionId
      * @param Subscription|null $subscription
@@ -59,6 +72,12 @@ class SubscriptionsController extends BaseController
         if ($subscription === null && $subscriptionId) {
             $subscription = Subscription::find()->anyStatus()->id($subscriptionId)->one();
         }
+
+        if (!$subscription) {
+            throw new NotFoundHttpException(Craft::t('commerce', 'Subscription not found'));
+        }
+
+        $this->enforceEditSubscriptionPermissions($subscription);
 
         $fieldLayout = Craft::$app->getFields()->getLayoutByType(Subscription::class);
 
@@ -118,6 +137,8 @@ class SubscriptionsController extends BaseController
             throw new NotFoundHttpException('Subscription not found');
         }
 
+        $this->enforceEditSubscriptionPermissions($subscription);
+
         $subscription->setFieldValuesFromRequest('fields');
 
         $subscription->setScenario(Element::SCENARIO_LIVE);
@@ -144,13 +165,14 @@ class SubscriptionsController extends BaseController
     public function actionRefreshPayments()
     {
         $this->requirePostRequest();
-        $this->requirePermission('commerce-manageSubscriptions');
 
         $subscriptionId = Craft::$app->getRequest()->getRequiredBodyParam('subscriptionId');
 
         if (!$subscription = Subscription::find()->anyStatus()->id($subscriptionId)->one()) {
             throw new NotFoundHttpException('Subscription not found');
         }
+
+        $this->enforceEditSubscriptionPermissions($subscription);
 
         $gateway = $subscription->getGateway();
         $gateway->refreshPaymentHistory($subscription);
@@ -270,11 +292,10 @@ class SubscriptionsController extends BaseController
         try {
             $subscriptionUid = $request->getValidatedBodyParam('subscriptionUid');
             $subscription = Subscription::find()->anyStatus()->uid($subscriptionUid)->one();
-            $userSession = Craft::$app->getUser();
 
             $validData = $subscriptionUid && $subscription;
             $validAction = $subscription->canReactivate();
-            $canModifySubscription = ($subscription->userId == $userSession->getId()) || $userSession->checkPermission('commerce-manageSubscriptions');
+            $canModifySubscription = $subscription->getIsEditable();
 
             if ($validData && $validAction && $canModifySubscription) {
                 if (!$plugin->getSubscriptions()->reactivateSubscription($subscription)) {
@@ -330,11 +351,10 @@ class SubscriptionsController extends BaseController
         try {
             $subscription = Subscription::find()->anyStatus()->uid($subscriptionUid)->one();
             $plan = Commerce::getInstance()->getPlans()->getPlanByUid($planUid);
-            $userSession = Craft::$app->getUser();
 
             $validData = $planUid && $plan && $subscriptionUid && $subscription;
             $validAction = $plan->canSwitchFrom($subscription->getPlan());
-            $canModifySubscription = ($subscription->userId == $userSession->getId()) || $userSession->checkPermission('commerce-manageSubscriptions');
+            $canModifySubscription = $subscription->getIsEditable();
 
             if ($validData && $validAction && $canModifySubscription) {
                 /** @var SubscriptionGateway $gateway */
@@ -402,12 +422,9 @@ class SubscriptionsController extends BaseController
 
         try {
             $subscriptionUid = $request->getValidatedBodyParam('subscriptionUid');
-
             $subscription = Subscription::find()->anyStatus()->uid($subscriptionUid)->one();
-            $userSession = Craft::$app->getUser();
-
             $validData = $subscriptionUid && $subscription;
-            $canModifySubscription = ($subscription->userId == $userSession->getId()) || $userSession->checkPermission('commerce-manageSubscriptions');
+            $canModifySubscription = $subscription->getIsEditable();
 
             if ($validData && $canModifySubscription) {
                 /** @var SubscriptionGateway $gateway */